fixed pdf and txt deletion when files local and not urls

+++ sentify/main.py
@@ -157,10 +157,12 @@ def pdf2text(pdf_or_stream, minline=4, trace=1):
 
             subprocess.run(["pdftotext", "-q", pdf, txt])
 
-            if exists_file(pdf):
-                remove_file(pdf)
             text = file2text(txt)
-            remove_file(txt)
+            if not isinstance(pdf_or_stream, str):
+                if exists_file(pdf):
+                    remove_file(pdf)
+                if exists_file(txt):
+                    remove_file(txt)
             if trace:
                 print("!!! pdftotext used")
         except FileNotFoundError:
